// web/default/src/features/business-preview/types.ts
/*
Copyright (C) 2023-2026 QuantumNous

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU Affero General Public License as
published by the Free Software Foundation, either version 3 of the
License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
GNU Affero General Public License for more details.

You should have received a copy of the GNU Affero General Public License
along with this program. If not, see <https://www.gnu.org/licenses/>.

For commercial licensing, please contact support@quantumnous.com
*/
import type { SubscriptionPlan } from '@/features/subscriptions/types'

export type PublicPaymentMethod = {
  type: string
  name: string
  min_topup: number
  recommended: boolean
}

export type PublicSubscriptionPlan = {
  id: number
  title: string
  subtitle: string
  price_amount: number
  currency: string
  billing_discount: number
  duration_unit: SubscriptionPlan['duration_unit']
  duration_value: number
  custom_seconds: number
  total_amount: number
  quota_reset_period: SubscriptionPlan['quota_reset_period']
  quota_reset_custom_seconds: number
  max_purchase_per_user: number
  unlimited: boolean
  non_refundable: boolean
}

export type BusinessPreviewData = {
  payment_methods: PublicPaymentMethod[]
  amount_options: number[]
  plans: PublicSubscriptionPlan[]
  quota_display_type: string
  currency_symbol: string
  purchase_requires_login: true
}

export type BusinessPreviewResponse = {
  success: boolean
  message?: string
  data?: BusinessPreviewData
}